requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
sqlalchemy
aiosqlite
alembic
